Fix command list-based devices on 64-bit platforms. Ghostscript always writes int values to the command list but old code interpreted the bit pattern as a long int, causing conversion of the negative values to large positive ones. Bug 689353.
DIFFERENCES:
None on 32-bit platforms.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8198 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gxclrast.c
@@ -89,10 +89,10 @@ cmd_print_bits(const byte * data, int width, int height, int raster)
 private long
 cmd_get_w(const byte * p, const byte ** rp)
 {
-    long val = *p++ & 0x7f;
+    int val = *p++ & 0x7f;
     int shift = 7;
 
-    for (; val |= (long)(*p & 0x7f) << shift, *p++ > 0x7f; shift += 7);
+    for (; val |= (int)(*p & 0x7f) << shift, *p++ > 0x7f; shift += 7);
     *rp = p;
     return val;
 }
